0.0.10 - Added UI elements

+++ EOY-Dystopia-Project-Main/Assets/Scripts/Scriptable Objects/GunData.cs
@@ -18,7 +18,7 @@ public class GunData : ScriptableObject
     public float cooldownRestPeriod;
     
     [Header("reloadConfig")]
-    public float magazineSize;
+    public int magazineSize;
     public float reloadTime;
     
     [Header("WeaponConfig")]
